+++ tibiawikisql/models/__init__.py
@@ -5,9 +5,9 @@ from tibiawikisql.models.charm import Charm
 from tibiawikisql.models.creature import Creature, CreatureAbility, CreatureDrop, CreatureMaxDamage
 from tibiawikisql.models.house import House
 from tibiawikisql.models.imbuement import Imbuement, ImbuementMaterial
-from tibiawikisql.models.item import Book, Item, ItemAttribute, ItemStoreOffer, Key
+from tibiawikisql.models.item import Book, Item, ItemAttribute, ItemProficiencyPerk, ItemStoreOffer, Key
 from tibiawikisql.models.mount import Mount
-from tibiawikisql.models.npc import Npc, NpcDestination, NpcOffer, NpcSpell, RashidPosition
+from tibiawikisql.models.npc import Npc, NpcDestination, NpcOffer, RashidPosition
 from tibiawikisql.models.outfit import Outfit, OutfitImage, OutfitQuest
 from tibiawikisql.models.quest import Quest, QuestDanger, QuestReward
 from tibiawikisql.models.spell import Spell